pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codebase-driven-agent"
version = "0.1.0"
description = "A general-purpose AI Agent platform driven by codebase. Currently focuses on problem analysis and error troubleshooting, with extensible architecture for future scenarios."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Codebase Driven Agent Team"}
]
keywords = ["ai", "agent", "codebase", "langchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=3.1.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "langchain>=1.2.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-community>=0.3.0",
    "sqlalchemy>=2.0.36",
    "pymysql>=1.1.2",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "gitpython>=3.1.45",
    "redis>=5.2.0",
    "sqlparse>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/your-org/codebase-driven-agent"
Documentation = "https://github.com/your-org/codebase-driven-agent/docs"
Repository = "https://github.com/your-org/codebase-driven-agent"

[tool.setuptools]
packages = ["codebase_driven_agent"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

